# internal/server/server.go
package server

import (
	"context"
	"fmt"
	"log/slog"
	"net"
	"net/http"
	"time"

	"github.com/allyourbase/ayb/internal/api"
	"github.com/allyourbase/ayb/internal/auth"
	"github.com/allyourbase/ayb/internal/config"
	"github.com/allyourbase/ayb/internal/httputil"
	"github.com/allyourbase/ayb/internal/realtime"
	"github.com/allyourbase/ayb/internal/schema"
	"github.com/allyourbase/ayb/internal/storage"
	"github.com/allyourbase/ayb/internal/webhooks"
	"github.com/allyourbase/ayb/openapi"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/jackc/pgx/v5/pgxpool"
)

// Server is the main HTTP server for AYB.
type Server struct {
	cfg                *config.Config
	router             *chi.Mux
	http               *http.Server
	logger             *slog.Logger
	schema             *schema.CacheHolder
	pool               *pgxpool.Pool
	authSvc            *auth.Service          // nil when auth disabled
	authRL             *auth.RateLimiter      // nil when auth disabled
	hub                *realtime.Hub
	webhookDispatcher  *webhooks.Dispatcher   // nil when pool is nil
	adminAuth          *adminAuth             // nil when admin.password not set
	startTime          time.Time
	logBuffer          *LogBuffer             // nil when not using buffered logging
}

// New creates a new Server with middleware and routes configured.
// authSvc and storageSvc may be nil when their features are disabled.
func New(cfg *config.Config, logger *slog.Logger, schemaCache *schema.CacheHolder, pool *pgxpool.Pool, authSvc *auth.Service, storageSvc *storage.Service) *Server {
	r := chi.NewRouter()

	// Global middleware (applies to all routes including admin SPA).
	r.Use(middleware.RequestID)
	r.Use(requestLogger(logger))
	r.Use(middleware.Recoverer)
	r.Use(corsMiddleware(cfg.Server.CORSAllowedOrigins))

	hub := realtime.NewHub(logger)

	// Webhooks (always created when pool is available).
	var webhookDispatcher *webhooks.Dispatcher
	if pool != nil {
		whStore := webhooks.NewStore(pool)
		webhookDispatcher = webhooks.NewDispatcher(whStore, logger)
		webhookDispatcher.SetDeliveryStore(whStore)
		webhookDispatcher.StartPruner(1*time.Hour, 7*24*time.Hour) // prune deliveries older than 7 days
	}

	s := &Server{
		cfg:               cfg,
		router:            r,
		logger:            logger,
		schema:            schemaCache,
		pool:              pool,
		authSvc:           authSvc,
		hub:               hub,
		webhookDispatcher: webhookDispatcher,
		startTime:         time.Now(),
	}
	if cfg.Admin.Password != "" {
		s.adminAuth = newAdminAuth(cfg.Admin.Password)
	}

	// Health check (no content-type restriction).
	r.Get("/health", s.handleHealth)

	// OpenAPI spec (no auth, no content-type restriction).
	r.Get("/api/openapi.yaml", handleOpenAPISpec)

	r.Route("/api", func(r chi.Router) {
		// Admin auth endpoints (no content-type enforcement — login needs JSON, status is GET).
		r.Get("/admin/status", s.handleAdminStatus)
		r.Post("/admin/auth", s.handleAdminLogin)

		// Admin SQL editor and RLS policy management (admin-auth gated, requires pool).
		if pool != nil {
			logger.Info("registering admin SQL and RLS routes")
			r.Route("/admin/sql", func(r chi.Router) {
				r.Use(s.requireAdminToken)
				r.Post("/", handleAdminSQL(pool))
			})

			// Admin RLS policy management.
			r.Route("/admin/rls", func(r chi.Router) {
				r.Use(s.requireAdminToken)
				r.Get("/", handleListRlsPolicies(pool))
				r.Post("/", handleCreateRlsPolicy(pool))
				r.Get("/{table}", handleListRlsPolicies(pool))
				r.Get("/{table}/status", handleGetRlsStatus(pool))
				r.Post("/{table}/enable", handleEnableRls(pool))
				r.Post("/{table}/disable", handleDisableRls(pool))
				r.Delete("/{table}/{policy}", handleDeleteRlsPolicy(pool))
			})
		} else {
			logger.Warn("pool is nil, skipping admin SQL and RLS routes")
		}

		// Admin user management (admin-auth gated, requires auth to be enabled).
		if authSvc != nil {
			r.Route("/admin/users", func(r chi.Router) {
				r.Use(s.requireAdminToken)
				r.Get("/", handleAdminListUsers(authSvc))
				r.Delete("/{id}", handleAdminDeleteUser(authSvc))
			})

			// Admin API key management.
			r.Route("/admin/api-keys", func(r chi.Router) {
				r.Use(s.requireAdminToken)
				r.Get("/", handleAdminListAPIKeys(authSvc))
				r.Post("/", handleAdminCreateAPIKey(authSvc))
				r.Delete("/{id}", handleAdminRevokeAPIKey(authSvc))
			})
		}

		// Admin logs (admin-auth gated).
		r.Route("/admin/logs", func(r chi.Router) {
			r.Use(s.requireAdminToken)
			r.Get("/", s.handleAdminLogs)
		})

		// Admin stats (admin-auth gated).
		r.Route("/admin/stats", func(r chi.Router) {
			r.Use(s.requireAdminToken)
			r.Get("/", s.handleAdminStats)
		})

		// Admin secrets management (admin-auth gated, requires auth service).
		if authSvc != nil {
			r.Route("/admin/secrets", func(r chi.Router) {
				r.Use(s.requireAdminToken)
				r.Post("/rotate", s.handleAdminSecretsRotate)
			})
		}

		// Storage routes accept multipart/form-data, mounted outside JSON content-type enforcement.
		if storageSvc != nil {
			storageHandler := storage.NewHandler(storageSvc, logger, cfg.Storage.MaxFileSizeBytes())
			r.Route("/storage", func(r chi.Router) {
				if authSvc != nil {
					r.Use(auth.OptionalAuth(authSvc))
				}
				r.Mount("/", storageHandler.Routes())
			})
		}

		// JSON API routes get content-type enforcement.
		r.Group(func(r chi.Router) {
			r.Use(middleware.AllowContentType("application/json"))

			// Auth endpoints (public, rate-limited).
			if authSvc != nil {
				authHandler := auth.NewHandler(authSvc, logger)
				// Configure OAuth providers from config.
				for name, p := range cfg.Auth.OAuth {
					if p.Enabled {
						authHandler.SetOAuthProvider(name, auth.OAuthClientConfig{
							ClientID:     p.ClientID,
							ClientSecret: p.ClientSecret,
						})
					}
				}
				if cfg.Auth.OAuthRedirectURL != "" {
					authHandler.SetOAuthRedirectURL(cfg.Auth.OAuthRedirectURL)
				}
				authHandler.SetOAuthPublisher(hub)
				rl := cfg.Auth.RateLimit
			if rl <= 0 {
				rl = 10
			}
			s.authRL = auth.NewRateLimiter(rl, time.Minute)
				r.Route("/auth", func(r chi.Router) {
					r.Use(s.authRL.Middleware)
					r.Mount("/", authHandler.Routes())
				})
			}

			r.Get("/schema", s.handleSchema)

			// Realtime SSE (handles its own auth for EventSource compatibility).
			rtHandler := realtime.NewHandler(hub, pool, authSvc, schemaCache, logger)
			r.Get("/realtime", rtHandler.ServeHTTP)

			// Webhook management (admin-only).
			if pool != nil {
				whStore := webhooks.NewStore(pool)
				whHandler := webhooks.NewHandler(whStore, whStore, logger)
				r.Route("/webhooks", func(r chi.Router) {
					r.Use(s.requireAdminToken)
					r.Mount("/", whHandler.Routes())
				})
			}

			// Mount auto-generated CRUD API.
			if pool != nil {
				apiHandler := api.NewHandler(pool, schemaCache, logger, hub, webhookDispatcher)
				if authSvc != nil {
					r.Group(func(r chi.Router) {
						// Accept either a valid admin HMAC token or a user JWT/API-key.
						r.Use(s.requireAdminOrUserAuth(authSvc))
						r.Mount("/", apiHandler.Routes())
					})
				} else {
					r.Mount("/", apiHandler.Routes())
				}
			}
		})
	})

	// Admin SPA (served from embedded UI assets).
	if cfg.Admin.Enabled {
		adminPath := cfg.Admin.Path
		if adminPath == "" {
			adminPath = "/admin"
		}
		spa := staticSPAHandler()
		// Mount under a Route group to avoid chi wildcard/redirect conflicts.
		r.Route(adminPath, func(sub chi.Router) {
			sub.Get("/", spa)
			sub.Get("/*", spa)
		})
	}

	return s
}

// SetLogBuffer attaches a log buffer for the /api/admin/logs endpoint.
func (s *Server) SetLogBuffer(lb *LogBuffer) {
	s.logBuffer = lb
}

// Router returns the chi router for registering additional routes.
func (s *Server) Router() *chi.Mux {
	return s.router
}

// Start begins listening for HTTP requests.
func (s *Server) Start() error {
	s.http = &http.Server{
		Addr:    s.cfg.Address(),
		Handler: s.router,
	}

	s.logger.Info("server starting", "address", s.cfg.Address())
	if err := s.http.ListenAndServe(); err != nil && err != http.ErrServerClosed {
		return fmt.Errorf("server error: %w", err)
	}
	return nil
}

// StartWithReady begins listening. It closes the ready channel once the
// listener is bound, then blocks serving requests.
func (s *Server) StartWithReady(ready chan<- struct{}) error {
	s.http = &http.Server{
		Addr:    s.cfg.Address(),
		Handler: s.router,
	}

	ln, err := net.Listen("tcp", s.cfg.Address())
	if err != nil {
		return fmt.Errorf("listen: %w", err)
	}

	s.logger.Info("server starting", "address", s.cfg.Address())
	close(ready)

	if err := s.http.Serve(ln); err != nil && err != http.ErrServerClosed {
		return fmt.Errorf("server error: %w", err)
	}
	return nil
}

// Shutdown gracefully stops the server.
func (s *Server) Shutdown(ctx context.Context) error {
	timeout := time.Duration(s.cfg.Server.ShutdownTimeout) * time.Second
	shutdownCtx, cancel := context.WithTimeout(ctx, timeout)
	defer cancel()

	s.logger.Info("shutting down server", "timeout", timeout)
	if s.authRL != nil {
		s.authRL.Stop()
	}
	if s.webhookDispatcher != nil {
		s.webhookDispatcher.Close()
	}
	s.hub.Close()
	return s.http.Shutdown(shutdownCtx)
}

func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	httputil.WriteJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func handleOpenAPISpec(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/yaml")
	w.Header().Set("Cache-Control", "public, max-age=3600")
	w.WriteHeader(http.StatusOK)
	w.Write(openapi.Spec)
}

func (s *Server) handleSchema(w http.ResponseWriter, r *http.Request) {
	sc := s.schema.Get()
	if sc == nil {
		httputil.WriteError(w, http.StatusServiceUnavailable, "schema cache not ready")
		return
	}
	httputil.WriteJSON(w, http.StatusOK, sc)
}