# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AllYourBase API
  description: |
    Backend-as-a-Service for PostgreSQL. Single binary. Auto-generated REST API
    with filtering, sorting, pagination, foreign key expansion, full-text search,
    batch operations, authentication, file storage, realtime events, and webhooks.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: AllYourBase
    url: https://allyourbase.io
servers:
  - url: http://localhost:8090
    description: Local development server

# Default: no auth required. Individual operations override with BearerAuth or AdminAuth.
security:
  - {}
...
paths:
...
  /api/admin/auth:
    post:
      tags: [Admin]
      summary: Admin login
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Admin auth not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/sql:
    post:
      tags: [Admin]
      summary: Execute SQL query
      description: Run arbitrary SQL against the database. Admin authentication required. 30-second timeout.
      operationId: adminSql
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SqlRequest"
      responses:
        "200":
          description: Query results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SqlResponse"
        "400":
          description: Invalid query or SQL error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List registered users
      description: |
        Paginated list of auth users with optional email search.
        Requires admin authentication and auth to be enabled.
      operationId: adminListUsers
      security:
        - AdminAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: perPage
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: search
          in: query
          description: Search by email (ILIKE match)
          schema:
            type: string
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a user
      description: Permanently deletes a user and all their sessions.
      operationId: adminDeleteUser
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User UUID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deleted
        "400":
          description: Missing user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/schema:
    get:
      tags: [Schema]
      summary: Get full database schema
      operationId: getSchema
      responses:
        "200":
          description: Database schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaCache"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Schema cache not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/realtime:
    get:
      tags: [Realtime]
      summary: Subscribe to realtime events (SSE)
      description: |
        Server-Sent Events stream for table changes. Events are filtered by
        RLS policies when authentication is enabled. Pass JWT token as query
        parameter for EventSource compatibility.
      operationId: realtimeSubscribe
      parameters:
        - name: tables
          in: query
          description: Comma-separated table names to subscribe to
          schema:
            type: string
            example: posts,comments
        - name: token
          in: query
          description: JWT token (for EventSource which cannot set headers)
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each event is JSON: {"action": "create"|"update"|"delete", "table": "name", "record": {...}}
        "401":
          description: Invalid or missing JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/storage/{bucket}:
    get:
      tags: [Storage]
      summary: List files in a bucket
      operationId: listFiles
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - name: prefix
          in: query
          description: Filter by filename prefix
          schema:
            type: string
        - name: limit
          in: query
          description: Max files to return
          schema:
            type: integer
        - name: offset
          in: query
          description: Number of files to skip
          schema:
            type: integer
      security:
        - BearerAuth: []
        - {}
      responses:
        "200":
          description: File list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageListResponse"
        "400":
          description: Invalid bucket name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Storage]
      summary: Upload a file
      operationId: uploadFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
              required:
                - file
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageObject"
        "400":
          description: Invalid form or file too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/storage/{bucket}/{name}:
    get:
      tags: [Storage]
      summary: Download a file
      operationId: downloadFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - $ref: "#/components/parameters/FileName"
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Storage]
      summary: Delete a file
      operationId: deleteFile
      parameters:
        - $ref: "#/components/parameters/BucketPath"
        - $ref: "#/components/parameters/FileName"
      security:
        - BearerAuth: []
        - {}
      responses:
        "204":
          description: File deleted
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/webhooks:
    get:
      tags: [Webhooks]
      summary: List webhooks
      operationId: listWebhooks
      security:
        - AdminAuth: []
      responses:
        "200":
          description: All webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Webhooks]
      summary: Create a webhook
      operationId: createWebhook
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid request (missing URL or invalid event)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/webhooks/{id}:
    get:
      tags: [Webhooks]
      summary: Get a webhook
      operationId: getWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "200":
          description: The webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Webhooks]
      summary: Update a webhook (partial)
      operationId: updateWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
      responses:
        "200":
          description: The updated webhook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          description: Invalid event type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Webhooks]
      summary: Delete a webhook
      operationId: deleteWebhook
      security:
        - AdminAuth: []
      parameters:
        - $ref: "#/components/parameters/WebhookId"
      responses:
        "204":
          description: Webhook deleted
        "401":
          description: Admin authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Webhook not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/login or /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin token from /api/admin/auth
  parameters:
...
    BucketPath:
      name: bucket
      in: path
      required: true
      description: Storage bucket name
      schema:
        type: string
    FileName:
      name: name
      in: path
      required: true
      description: File name within the bucket
      schema:
        type: string
    WebhookId:
      name: id
      in: path
      required: true
      description: Webhook UUID
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "collection not found: nonexistent"
        data:
          type: object
          description: Field-level validation details
          additionalProperties: true
...
    StorageObject:
      type: object
      required: [id, bucket, name, size, contentType, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        bucket:
          type: string
        name:
          type: string
        size:
          type: integer
          format: int64
          description: File size in bytes
        contentType:
          type: string
          example: image/png
        userId:
          type: string
          format: uuid
          description: Uploader's user ID (if authenticated)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StorageListResponse:
      type: object
      required: [items, totalItems]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StorageObject"
        totalItems:
          type: integer
...
    WebhookRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: URL to receive webhook events
        secret:
          type: string
          description: HMAC-SHA256 signing secret for payload verification
        events:
          type: array
          items:
            type: string
            enum: [create, update, delete]
          description: "Event types to subscribe to (default: all)"
          default: [create, update, delete]
        tables:
          type: array
          items:
            type: string
          description: "Table names to filter (default: all tables)"
          default: []
        enabled:
          type: boolean
          description: Whether the webhook is active
          default: true
    WebhookResponse:
      type: object
      required: [id, url, hasSecret, events, tables, enabled, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        hasSecret:
          type: boolean
          description: Whether a signing secret is configured (secret value is never exposed)
        events:
          type: array
          items:
            type: string
            enum: [create, update, delete]
        tables:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    SqlRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          description: SQL query to execute
          example: "SELECT * FROM posts LIMIT 10"
    SqlResponse:
      type: object
      required: [columns, rows, rowCount, durationMs]
      properties:
        columns:
          type: array
          items:
            type: string
          description: Column names from the result set
        rows:
          type: array
          items:
            type: array
            items: {}
          description: Row data as arrays of values
        rowCount:
          type: integer
          description: Number of rows returned
        durationMs:
          type: integer
          format: int64
          description: Query execution time in milliseconds
    AdminLoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminLoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Admin Bearer token (HMAC-derived, valid for server lifetime)
...
    AdminUser:
      type: object
      required: [id, email, emailVerified, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserListResponse:
      type: object
      required: [items, page, perPage, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        page:
          type: integer
        perPage:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
    SchemaCache:
      type: object
      properties:
        tables:
          type: object
          description: "Map of schema-qualified table names to table definitions"
          additionalProperties:
            $ref: "#/components/schemas/SchemaTable"
        functions:
          type: object
          description: "Map of schema-qualified function names to function definitions"
          additionalProperties:
            $ref: "#/components/schemas/SchemaFunction"
        schemas:
          type: array
          items:
            type: string
          description: List of database schemas
        builtAt:
          type: string
          format: date-time
    SchemaTable:
      type: object
      properties:
        schema:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [table, view, materialized_view, partitioned_table]
        comment:
          type: string
        columns:
          type: array
          items:
            $ref: "#/components/schemas/SchemaColumn"
        primaryKey:
          type: array
          items:
            type: string
        foreignKeys:
          type: array
          items:
            $ref: "#/components/schemas/SchemaForeignKey"
        relationships:
          type: array
          items:
            $ref: "#/components/schemas/SchemaRelationship"
    SchemaColumn:
      type: object
      properties:
        name:
          type: string
        position:
          type: integer
        type:
          type: string
          description: PostgreSQL type name
        nullable:
          type: boolean
        default:
          type: string
        comment:
          type: string
        isPrimaryKey:
          type: boolean
        jsonType:
          type: string
          description: Mapped JSON type (string, number, boolean, object, array)
        enumValues:
          type: array
          items:
            type: string
    SchemaForeignKey:
      type: object
      properties:
        constraintName:
          type: string
        columns:
          type: array
          items:
            type: string
        referencedSchema:
          type: string
        referencedTable:
          type: string
        referencedColumns:
          type: array
          items:
            type: string
        onUpdate:
          type: string
        onDelete:
          type: string
    SchemaRelationship:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [many-to-one, one-to-many]
        fromSchema:
          type: string
        fromTable:
          type: string
        fromColumns:
          type: array
          items:
            type: string
        toSchema:
          type: string
        toTable:
          type: string
        toColumns:
          type: array
          items:
            type: string
        fieldName:
          type: string
    SchemaFunction:
      type: object
      properties:
        schema:
          type: string
        name:
          type: string
        comment:
          type: string
        parameters:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              position:
                type: integer
        returnType:
          type: string
        returnsSet:
          type: boolean